Switching to Spring Data Repositories

+++ src/main/java/com/cezarykluczynski/carmen/admin/web/endpoint/dto/LanguagesListUpdateStatusDTO.java
@@ -15,8 +15,8 @@ public class LanguagesListUpdateStatusDTO {
 
     private Boolean updatable;
 
-    private Integer persistedLanguagesCount;
+    private Long persistedLanguagesCount;
 
-    private Integer linguistLanguagesCount;
+    private Long linguistLanguagesCount;
 
 }
